+++ sq/CHANGELOG.md
@@ -1,3 +1,12 @@
+# Version 0.0.2
+
+* Added `FromJSON`, `ToJSON` instances for `Name`.
+
+* Improved type-parameter order in `zero`, `one`, `foldM` and similar.
+
+* Added manual transactional migrations suport.
+
+
 # Version 0.0.1
 
 * Initial version.